+++ src/data/school_holidays.fr.js
@@ -6,7 +6,6 @@ const xml = require('fast-xml-parser');
 
 const Utils = require('../utils');
 
-
 function close() {
   // Clear the cache
   memoize.clear(getHolidays);
@@ -18,23 +17,25 @@ function initialize() {
 }
 
 function isHolidays(date, region) {
-  return getHolidays().then((holidays) => {
-    const years = holidays[region];
+  return getHolidays()
+    .then((holidays) => {
+      const years = holidays[region];
 
-    if (!years)
-      throw new RangeError(`No school holidays information for the region "${region}".`);
+      if (!years) {
+        throw new RangeError(`No school holidays information for the region "${region}".`);
+      }
 
-    const months = years[date.year];
+      const months = years[date.year];
 
-    if (!months)
-      throw new RangeError(`No school holidays information for the region "${region}" in ${date.year}.`);
+      if (!months) {
+        throw new RangeError(`No school holidays information for the region "${region}" in ${date.year}.`);
+      }
 
-    const month = months[date.month];
+      const month = months[date.month];
 
-    return month && month[date.day];
-  });
+      return month && month[date.day];
+    });
 }
-
 
 function fetchHolidays() {
   return fetch('https://telechargement.index-education.com/vacances.xml')
@@ -44,16 +45,22 @@ function fetchHolidays() {
 }
 
 function formatHolidays(data) {
-  return [].concat(...Object.values(data)).reduce(reduceHolidays, {});
+  return [].concat(...Object.values(data))
+    .reduce(reduceHolidays, {});
 }
 
 function indexHolidays(years, current) {
-  const start = parseDate(current.debut).plus(oneDay);
+  const start = parseDate(current.debut)
+    .plus(oneDay);
   const year = start.month <= 7 ? start.year - 1 : start.year;
   const value = { start, end: parseDate(current.fin) };
 
-  if (year in years) years[year].push(value);
-  else years[year] = [value];
+  if (year in years) {
+    years[year].push(value);
+  }
+  else {
+    years[year] = [value];
+  }
 
   return years;
 }
@@ -68,22 +75,28 @@ function parse(text) {
   const data = patch(xml.parse(text, PARSER_OPTIONS).root);
   const zones = data.calendrier.zone.reduce(indexZones, {});
 
-  const exceptions = toArray(data.exceptionszones.exceptionzone).reduce((exceptions, exception) => {
-    const academyId = exception.academie;
-    const year = exception.annee;
-    const holidays = zones[exception.zone][year];
+  const exceptions = toArray(data.exceptionszones.exceptionzone)
+    .reduce((exceptions, exception) => {
+      const academyId = exception.academie;
+      const year = exception.annee;
+      const holidays = zones[exception.zone][year];
 
-    /* istanbul ignore if */
-    if (academyId in exceptions) exceptions[academyId][year] = holidays;
-    else exceptions[academyId] = { [year]: holidays };
+      /* istanbul ignore if */
+      if (academyId in exceptions) {
+        exceptions[academyId][year] = holidays;
+      }
+      else {
+        exceptions[academyId] = { [year]: holidays };
+      }
 
-    return exceptions;
-  }, {});
+      return exceptions;
+    }, {});
 
   const regions = data.academies.academie.reduce((regions, academy) => {
     const holidays = formatHolidays({ ...zones[academy.zone], ...exceptions[academy.libelle] });
 
-    toArray(academy.departement).forEach((region) => regions[region.numero] = holidays);
+    toArray(academy.departement)
+      .forEach((region) => regions[region.numero] = holidays);
 
     return regions;
   }, {});
@@ -104,22 +117,26 @@ function patch(data) {
 
   return data;
 
-
   function addRegionTo(academies, academyId, regionId) {
     const academy = findAcademy(academies, academyId);
     const regions = toArray(academy.departement);
 
     /* istanbul ignore if */
-    if (regions.some((region) => region.numero == regionId))
+    if (regions.some((region) => region.numero == regionId)) {
       console.warn(`Adding region "${regionId}" to ${academyId} academy is not needed anymore.`);
-    else academy.departement = regions.concat([{ numero: regionId }]);
+    }
+    else {
+      academy.departement = regions.concat([{ numero: regionId }]);
+    }
   }
 
   function findAcademy(academies, academyId) {
     const academy = academies.find((academy) => academy.libelle === academyId);
 
     /* istanbul ignore if */
-    if (!academy) throw new Error(`Unable to patch ${academyId} academy.`);
+    if (!academy) {
+      throw new Error(`Unable to patch ${academyId} academy.`);
+    }
 
     return academy;
   }
@@ -130,15 +147,19 @@ function patch(data) {
     const filtered = regions.filter((region) => region.numero != regionId);
 
     /* istanbul ignore if */
-    if (regions.length === filtered.length)
+    if (regions.length === filtered.length) {
       console.warn(`Removing region "${regionId}" from ${academyId} academy is not needed anymore.`);
-    else academy.departement = filtered;
+    }
+    else {
+      academy.departement = filtered;
+    }
   }
 }
 
 function reduceHolidays(years, holidays) {
   const start = holidays.start;
-  const days = Math.floor(holidays.end.diff(start).as('days'));
+  const days = Math.floor(holidays.end.diff(start)
+    .as('days'));
 
   for (let i = 0; i < days; i++) {
     const date = start.plus({ days: i });
@@ -149,9 +170,16 @@ function reduceHolidays(years, holidays) {
     if (year in years) {
       const months = years[year];
 
-      if (month in months) months[month][day] = true;
-      else months[month] = { [day]: true };
-    } else years[year] = { [month]: { [day]: true } };
+      if (month in months) {
+        months[month][day] = true;
+      }
+      else {
+        months[month] = { [day]: true };
+      }
+    }
+    else {
+      years[year] = { [month]: { [day]: true } };
+    }
   }
 
   return years;
@@ -165,7 +193,6 @@ function toArray(value) {
   return Array.isArray(value) ? value : [value];
 }
 
-
 const MEMOIZE_OPTIONS = { maxAge: 2 * 24 * 3600 * 1000 };
 const RETRY_OPTIONS = { retries: 5, minTimeout: 100 };
 const PARSER_OPTIONS = { attributeNamePrefix: '', ignoreAttributes: false, parseNodeValue: false };
@@ -173,9 +200,8 @@ const DateTime = luxon.DateTime;
 const getHolidays = memoize(() => retry(fetchHolidays, RETRY_OPTIONS), MEMOIZE_OPTIONS);
 const oneDay = { day: 1 };
 
-
 module.exports = {
   close,
   initialize,
-  isHolidays,
+  isHolidays
 };