+++ src/endpoint/destroy.js
@@ -16,5 +16,4 @@ async function destroy() {
     });
 }
 
-
 module.exports = destroy;