+++ src/providers/public_holiday.js
@@ -4,13 +4,13 @@ const memoize = require('mem');
 
 const Constants = require('../constants');
 
-
 async function initialize(provider) {
   const options = provider.options;
   const country = options.country;
 
-  if (typeof country !== 'string')
+  if (typeof country !== 'string') {
     throw new TypeError(`The "country" option of the public holiday provider must be a "string". Received "${typeof country}".`);
+  }
 
   const context = provider.context;
 
@@ -23,9 +23,11 @@ async function initialize(provider) {
     context.fixed = fixed;
     context.easterOffseted = easterOffseted;
     context.regions = holidays.regions ? formatRegions(holidays.regions, fixed, easterOffseted) : /* istanbul ignore next */{};
-  } catch (error)/* istanbul ignore next */ {
-    if (error.code === 'MODULE_NOT_FOUND')
+  }
+  catch (error)/* istanbul ignore next */ {
+    if (error.code === 'MODULE_NOT_FOUND') {
       throw new RangeError(`The "country" option of the public holiday provider must be valid. Received "${country}".`);
+    }
 
     throw error;
   }
@@ -36,7 +38,6 @@ async function initialize(provider) {
   context.cache = cache;
   context.easter = memoize(getEasterDate, { cache });
   provider.refresh.period = 24 * 3600;
-
 
   function getEasterDate(year) {
     const date = easter(year);
@@ -49,13 +50,13 @@ async function extendConfiguration() {
   // TODO: Check the endpoint's metadata
 
   return {
-    [HOLIDAY]: { type: 'enum' },
+    [HOLIDAY]: { type: 'enum' }
   };
 }
 
 async function extendRecord(endpoint, record) {
   return {
-    [HOLIDAY]: isHoliday.call(this, record[PARSED_RECORD][DATE], endpoint.metadata.region) ? 'YES' : 'NO',
+    [HOLIDAY]: isHoliday.call(this, record[PARSED_RECORD][DATE], endpoint.metadata.region) ? 'YES' : 'NO'
   };
 }
 
@@ -63,7 +64,6 @@ async function close() {
   // Clear the cache
   this.context.cache.clear();
 }
-
 
 function formatRegions(regions, fixed, easterOffseted) {
   return regions.reduce((regions, element) => {
@@ -75,7 +75,7 @@ function formatRegions(regions, fixed, easterOffseted) {
         : days),
       easterOffseted: holidays.easterOffseted
         ? Object.assign(indexArray(holidays.easterOffseted), easterOffseted)
-        : easterOffseted,
+        : easterOffseted
     };
 
     return Object.assign(regions, element[0].reduce(indexWith(regionalHolidays), {}));
@@ -103,15 +103,17 @@ function indexWith(value) {
 function isHoliday(date, region) {
   const holidays = getRegionalHolidays.call(this, region);
 
-  if (holidays.fixed[date.month][date.day]) return true;
+  if (holidays.fixed[date.month][date.day]) {
+    return true;
+  }
 
   const easterDate = this.context.easter(date.year);
   // Comparison between easter and the current date needs to be done in the same timezone
-  const easterOffset = Math.floor(date.setZone(UTC_ZONE, KEEP_LOCAL_TIME).diff(easterDate, 'days').days);
+  const easterOffset = Math.floor(date.setZone(UTC_ZONE, KEEP_LOCAL_TIME)
+    .diff(easterDate, 'days').days);
 
   return holidays.easterOffseted[easterOffset];
 }
-
 
 const PARSED_RECORD = Constants.PARSED_RECORD;
 const DATE = Constants.DATE_FEATURE;
@@ -122,11 +124,10 @@ const HOLIDAY = 'public_holiday';
 const DateTime = luxon.DateTime;
 const Indexer = indexWith(true);
 
-
 module.exports = {
   HOLIDAY,
   close,
   extendConfiguration,
   extendRecord,
-  initialize,
+  initialize
 };