+++ src/providers/school_holidays.js
@@ -1,20 +1,22 @@
 const Constants = require('../constants');
 
-
 async function initialize(provider) {
   const options = provider.options;
   const country = options.country;
 
-  if (typeof country !== 'string')
+  if (typeof country !== 'string') {
     throw new TypeError(`The "country" option of the school holidays provider must be a "string". Received "${typeof country}".`);
+  }
 
   const context = provider.context;
 
   try {
     context.holidays = require(`../data/school_holidays.${country}`);
-  } catch (error)/* istanbul ignore next */ {
-    if (error.code === 'MODULE_NOT_FOUND')
+  }
+  catch (error)/* istanbul ignore next */ {
+    if (error.code === 'MODULE_NOT_FOUND') {
       throw new RangeError(`The "country" option of the school holidays provider must be valid. Received "${country}".`);
+    }
 
     throw error;
   }
@@ -28,7 +30,7 @@ async function extendConfiguration() {
   // TODO: Check the endpoint's metadata
 
   return {
-    [HOLIDAY]: { type: 'enum' },
+    [HOLIDAY]: { type: 'enum' }
   };
 }
 
@@ -47,11 +49,9 @@ async function close() {
   return this.context.holidays.close();
 }
 
-
 function formatExtension(isHolidays) {
-  return { [HOLIDAY]: isHolidays ? 'YES' : 'NO', };
+  return { [HOLIDAY]: isHolidays ? 'YES' : 'NO' };
 }
-
 
 const PARSED_RECORD = Constants.PARSED_RECORD;
 const DATE = Constants.DATE_FEATURE;
@@ -59,11 +59,10 @@ const DATE = Constants.DATE_FEATURE;
 const HOLIDAY = 'school_holiday';
 const UNKNOWN = { [HOLIDAY]: 'UNKNOWN' };
 
-
 module.exports = {
   HOLIDAY,
   close,
   extendConfiguration,
   extendRecord,
-  initialize,
+  initialize
 };