+++ test/endpoint/destroy.test.js
@@ -2,11 +2,9 @@ const test = require('ava');
 
 const Helpers = require('../helpers');
 
-
 test.before(require('dotenv').config);
 test.beforeEach(Helpers.createEndpointContext);
 test.afterEach.always(Helpers.destroyEndpointContext);
-
 
 test('destroys the endpoint\'s related agent', (t) => {
   const context = t.context;