+++ CHANGELOG.md
@@ -10,7 +10,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ### Fixed
 
-- Enable `roundDate` to have period equal to NaN 
+- Enable `roundDate` to have period equal to NaN
 
 ## [0.4.1](https://github.com/craft-ai/craft-ai-kit-energy/compare/v0.4.0...v0.4.1) - 2019-10-23 ##